Safer access to constructing the filelist of produced tuples
Now explicitely checks for the production time-tag (YYMMDD_HHMMSS) and take the most recent one.
In this way we avoid having the same events twice in case a sample has been produced two times (without changing the productionLabel).
The glob statements are also simplified given that we moved in to a consistent directory structure (regardless if a sample was produced by crab or locally)
Some explicit checks and warnings added.